v43 review-nits: doc + format-spec cleanups
Three cosmetic fixes from code review of feat/v43-audit-cleanup
(no behavior change, no new findings):

- homekit.cpp:985 — W36 doc: parameter order in the explanatory
  comment was `(eraseap, wifioff)`; arduino-esp32's actual signature
  is `WiFi.disconnect(wifioff, eraseap, timeoutLength)`.
- ratgdo.cpp:456 — W38 doc: cited line for the elapsed-time gate
  was homekit.cpp:1010; actual line at HEAD is homekit.cpp:1026.
- web.cpp:1440 — W35 nit: `snprintf` format spec used `%u` for
  WiFi.channel() which returns int32_t. Changed to `%d` with
  matching `(int)` cast for sign-correctness; -Wno-format had been
  masking the type mismatch.

+++ src/web.cpp
@@ -1437,7 +1437,7 @@ void build_status_json(char *json)
         strncpy(ssidStr,  s_ssid.c_str(),  sizeof(ssidStr)  - 1); ssidStr[sizeof(ssidStr)   - 1] = '\0';
         strncpy(bssidStr, s_bssid.c_str(), sizeof(bssidStr) - 1); bssidStr[sizeof(bssidStr) - 1] = '\0';
     }
-    snprintf(rssiBuf, sizeof(rssiBuf), "%d dBm, Channel %u", (int)WiFi.RSSI(), WiFi.channel());
+    snprintf(rssiBuf, sizeof(rssiBuf), "%d dBm, Channel %d", (int)WiFi.RSSI(), (int)WiFi.channel());
     JSON_ADD_STR("macAddress", macStr);
     JSON_ADD_STR("wifiSSID", ssidStr);
     JSON_ADD_STR("wifiRSSI", rssiBuf);
